+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Added
 - Added offline, online, resync method to networks with failed heartbeat.
+- Heartbeat now checks also if node is added to the node list and if not, reports to user.
 - Configuration for heartbeat interval (`heartbeat_interval`, default 30 mins).
-- Websocket server for real time push messages to your browser! Running by default on port 40000, if not available, will test next possible port until reaches port 40200. Check the assigned port from logs! **NOTE: REMEMBER TO FORWARD PORT TO YOUR SERVER IF YOU ARE BEHIND A NAT!**
+- Websocket server for real time push messages to your browser. To enable, set `websocket_port` in configuration file.
 
 ### Changed
 - Default interval for heartbeat thread is now 30 minutes for allowing node to try resync method.